change testFixtures package
