+++ jpa-spec-kotlin-dsl-hibernate/src/test/kotlin/io/github/alfonsoristorato/jpaspeckotlindslhibernate/archunit/ArchUnitRulesRunner.kt
@@ -1,5 +1,5 @@
 package io.github.alfonsoristorato.jpaspeckotlindslhibernate.archunit
 
-import io.github.alfonsoristorato.jpaspeckotlindsl.archunit.BaseArchUnitRulesRunner
+import io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.archunit.BaseArchUnitRulesRunner
 
 class ArchUnitRulesRunner : BaseArchUnitRulesRunner("io.github.alfonsoristorato.jpaspeckotlindslhibernate")
