+++ jpa-spec-kotlin-dsl/src/test/kotlin/io/github/alfonsoristorato/jpaspeckotlindsl/archunit/ArchUnitRulesRunner.kt
@@ -1,3 +1,5 @@
 package io.github.alfonsoristorato.jpaspeckotlindsl.archunit
 
+import io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.archunit.BaseArchUnitRulesRunner
+
 class ArchUnitRulesRunner : BaseArchUnitRulesRunner("io.github.alfonsoristorato.jpaspeckotlindsl")
