+++ jpa-spec-kotlin-dsl/src/test/kotlin/io/github/alfonsoristorato/jpaspeckotlindsl/specification/combiner/CombinerTest.kt
@@ -1,12 +1,12 @@
 package io.github.alfonsoristorato.jpaspeckotlindsl.specification.combiner
 
-import io.github.alfonsoristorato.jpaspeckotlindsl.jpasetup.entity.Persona
-import io.github.alfonsoristorato.jpaspeckotlindsl.jpasetup.repository.PersonaRepository
-import io.github.alfonsoristorato.jpaspeckotlindsl.jpasetup.testconfig.SpringBootTestEnhanced
 import io.github.alfonsoristorato.jpaspeckotlindsl.specification.comparison.greaterThan
 import io.github.alfonsoristorato.jpaspeckotlindsl.specification.comparison.lessThan
 import io.github.alfonsoristorato.jpaspeckotlindsl.specification.equality.equal
-import io.github.alfonsoristorato.jpaspeckotlindsl.util.TestFixtures
+import io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.jpasetup.entity.Persona
+import io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.jpasetup.repository.PersonaRepository
+import io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.jpasetup.testconfig.SpringBootTestEnhanced
+import io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.util.TestFixtures
 import io.kotest.assertions.assertSoftly
 import io.kotest.core.spec.style.ExpectSpec
 import io.kotest.matchers.collections.shouldHaveSize
