+++ jpa-spec-kotlin-dsl/src/testFixtures/kotlin/io/github/alfonsoristorato/jpaspeckotlindsl/testfixtures/jpasetup/entity/Comment.kt
@@ -1,27 +1,26 @@
-package io.github.alfonsoristorato.jpaspeckotlindsl.jpasetup.entity
+package io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.jpasetup.entity
 
-import jakarta.persistence.ElementCollection
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.ManyToOne
+import jakarta.persistence.OneToOne
 import jakarta.persistence.SequenceGenerator
 import jakarta.persistence.Table
 
 @Entity
 @Table
-class Post(
+class Comment(
     @ManyToOne(fetch = FetchType.LAZY)
+    val post: Post,
+    @OneToOne(fetch = FetchType.LAZY)
     val persona: Persona,
-    val title: String,
     val content: String,
-    @ElementCollection
-    val tags: Set<String>,
 ) {
     @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "post_id_sequence")
-    @SequenceGenerator(name = "post_id_sequence", allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "comment_id_sequence")
+    @SequenceGenerator(name = "comment_id_sequence", allocationSize = 1)
     var id: Long? = null
 }
