+++ jpa-spec-kotlin-dsl/src/testFixtures/kotlin/io/github/alfonsoristorato/jpaspeckotlindsl/testfixtures/jpasetup/repository/PersonaRepository.kt
@@ -1,6 +1,6 @@
-package io.github.alfonsoristorato.jpaspeckotlindsl.jpasetup.repository
+package io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.jpasetup.repository
 
-import io.github.alfonsoristorato.jpaspeckotlindsl.jpasetup.entity.Persona
+import io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.jpasetup.entity.Persona
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.stereotype.Repository
