+++ jpa-spec-kotlin-dsl/src/testFixtures/kotlin/io/github/alfonsoristorato/jpaspeckotlindsl/testfixtures/jpasetup/testconfig/SpringBootTestEnhanced.kt
@@ -1,6 +1,6 @@
-package io.github.alfonsoristorato.jpaspeckotlindsl.jpasetup.testconfig
+package io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.jpasetup.testconfig
 
-import io.github.alfonsoristorato.jpaspeckotlindsl.jpasetup.JpaSetupApplication
+import io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.jpasetup.JpaSetupApplication
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.context.annotation.Import
 import org.springframework.transaction.annotation.Transactional
