+++ jpa-spec-kotlin-dsl/src/testFixtures/resources/kotest.properties
@@ -1,1 +1,1 @@
-kotest.framework.config.fqn=io.github.alfonsoristorato.jpaspeckotlindsl.kotest.KotestProjectConfig
+kotest.framework.config.fqn=io.github.alfonsoristorato.jpaspeckotlindsl.testfixtures.kotest.KotestProjectConfig
